+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # main (unreleased)
+- [feature]: Add the ability to perform the check in the specified path
 
 # 0.2.0
 - [[feature]: add links to the GitHub repo to be able to navigate from the npm package page](https://github.com/upgradejs/depngn/pull/27)